refactor(service): ♻️ 优化MyBatis-Plus相关代码及依赖升级
将MyBatis-Plus从3.2.0升级到3.5.10版本，并替换mybatis-plus-join为mybatis-plus-join-boot-starter 1.5.3版本。同时将druid升级至1.2.24版本并使用spring-boot-starter方式引入。

主要变更包括：
1. 修改查询返回类型从Integer改为Long以兼容MyBatis-Plus新版本API
2. 更新LambdaQueryChainWrapper的导入路径
3. 为实体类添加@TableId注解
4. 优化分页查询中的count类型
5. 统一命令进度字段类型为Long

这些修改提高了代码与新版本MyBatis-Plus的兼容性，并遵循了最新的最佳实践。

+++ datasophon-infrastructure/src/main/java/com/datasophon/dao/entity/ClusterServiceInstanceEntity.java
@@ -17,17 +17,15 @@
 
 package com.datasophon.dao.entity;
 
-import com.datasophon.dao.enums.NeedRestart;
-import com.datasophon.dao.enums.ServiceState;
-
-import java.io.Serializable;
-import java.util.Date;
-
-import lombok.Data;
-
 import com.baomidou.mybatisplus.annotation.TableField;
 import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
+import com.datasophon.dao.enums.NeedRestart;
+import com.datasophon.dao.enums.ServiceState;
+import lombok.Data;
+
+import java.io.Serializable;
+import java.util.Date;
 
 @TableName("t_ddh_cluster_service_instance")
 @Data
@@ -74,7 +72,7 @@ public class ClusterServiceInstanceEntity implements Serializable {
     private String dashboardUrl;
 
     @TableField(exist = false)
-    private Integer alertNum;
+    private Long alertNum;
 
     private Integer sortNum;
 }
