+++ datasophon-service/src/main/java/com/datasophon/api/service/host/dto/QueryHostListPageDTO.java
@@ -70,5 +70,5 @@ public class QueryHostListPageDTO {
 
     private String nodeLabel;
 
-    private Integer serviceRoleNum;
+    private Long serviceRoleNum;
 }
