+++ datasophon-service/src/main/java/com/datasophon/api/service/impl/AlertGroupServiceImpl.java
@@ -18,7 +18,7 @@
 package com.datasophon.api.service.impl;
 
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
-import com.baomidou.mybatisplus.extension.service.additional.query.impl.LambdaQueryChainWrapper;
+import com.baomidou.mybatisplus.extension.conditions.query.LambdaQueryChainWrapper;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import com.datasophon.api.service.AlertGroupService;
 import com.datasophon.api.service.ClusterAlertGroupMapService;
@@ -67,7 +67,7 @@ public class AlertGroupServiceImpl extends ServiceImpl<AlertGroupMapper, AlertGr
         LambdaQueryChainWrapper<AlertGroupEntity> wrapper = this.lambdaQuery()
                 .in(AlertGroupEntity::getId, groupIds)
                 .like(StringUtils.isNotBlank(alertGroupName), AlertGroupEntity::getAlertGroupName, alertGroupName);
-        int count = wrapper.count() == null ? 0 : wrapper.count();
+        long count = wrapper.count() == null ? 0 : wrapper.count();
         List<AlertGroupEntity> alertGroupList = wrapper.last("limit " + offset + "," + pageSize).list();
         if (CollectionUtils.isEmpty(alertGroupList)) {
             return Result.successEmptyCount();
