+++ datasophon-service/src/main/java/com/datasophon/api/service/impl/ClusterGroupServiceImpl.java
@@ -220,7 +220,7 @@ public class ClusterGroupServiceImpl extends ServiceImpl<ClusterGroupMapper, Clu
     @Override
     public Result deleteUserGroup(Integer id) {
         ClusterGroup clusterGroup = this.getById(id);
-        Integer num = userGroupService.countGroupUserNum(id);
+        long num = userGroupService.countGroupUserNum(id);
         if (num > 0) {
             return Result.error(Status.USER_GROUP_TIPS_ONE.getMsg());
         }
@@ -249,7 +249,7 @@ public class ClusterGroupServiceImpl extends ServiceImpl<ClusterGroupMapper, Clu
     @Override
     public Result deleteUserGroupOnK8s(Integer id) {
         ClusterGroup clusterGroup = this.getById(id);
-        Integer num = userGroupService.countGroupUserNum(id);
+        long num = userGroupService.countGroupUserNum(id);
         if (num > 0) {
             return Result.error(Status.USER_GROUP_TIPS_ONE.getMsg());
         }
@@ -283,7 +283,7 @@ public class ClusterGroupServiceImpl extends ServiceImpl<ClusterGroupMapper, Clu
                 clusterGroup.setClusterUsers(clusterUsers);
             }
         }
-        int total = this.count(new QueryWrapper<ClusterGroup>().
+        long total = this.count(new QueryWrapper<ClusterGroup>().
                 like(StringUtils.isNotBlank(groupName), Constants.GROUP_NAME, groupName)
                 .eq(Constants.CLUSTER_ID, clusterId));
         return Result.success(list).put(Constants.TOTAL, total);
