+++ datasophon-service/src/main/java/com/datasophon/api/service/impl/ClusterServiceCommandServiceImpl.java
@@ -189,7 +189,7 @@ public class ClusterServiceCommandServiceImpl
         List<ClusterServiceCommandEntity> list = this.list(new QueryWrapper<ClusterServiceCommandEntity>()
                 .eq(Constants.CLUSTER_ID, clusterId)
                 .orderByDesc(Constants.CREATE_TIME).last("limit " + offset + "," + pageSize));
-        Integer total = this.count(new QueryWrapper<ClusterServiceCommandEntity>()
+        long total = this.count(new QueryWrapper<ClusterServiceCommandEntity>()
                 .eq(Constants.CLUSTER_ID, clusterId));
         for (ClusterServiceCommandEntity commandEntity : list) {
             commandEntity.setCommandStateCode(commandEntity.getCommandState().getValue());
