+++ datasophon-service/src/main/java/com/datasophon/api/service/impl/UserInfoServiceImpl.java
@@ -119,11 +119,11 @@ public class UserInfoServiceImpl extends ServiceImpl<UserInfoMapper, UserInfoEnt
 
     @Override
     public Result getUserListByPage(String username, Integer page, Integer pageSize) {
-        Integer offset = (page - 1) * pageSize;
+        int offset = (page - 1) * pageSize;
         List<UserInfoEntity> list = this.list(
                 new QueryWrapper<UserInfoEntity>().like(StringUtils.isNotBlank(username), Constants.USERNAME, username)
                         .last("limit " + offset + "," + pageSize));
-        int total = this.count(new QueryWrapper<UserInfoEntity>().like(StringUtils.isNotBlank(username),
+        long total = this.count(new QueryWrapper<UserInfoEntity>().like(StringUtils.isNotBlank(username),
                 Constants.USERNAME, username));
         return Result.success().put(Constants.DATA, list).put(Constants.TOTAL, total);
     }
